+++ src/LieGroups.jl
@@ -40,7 +40,8 @@ using ManifoldsBase:
     get_parameter,
     internal_value,
     submanifold_component,
-    submanifold_components
+    submanifold_components,
+    tangent_vector_type
 using StaticArrays
 #
 #
@@ -50,6 +51,7 @@ using StaticArrays
 import Manifolds: apply, apply!, compose, identity_element, is_identity
 # Both define the following structs, so these for now lead to asking for explicit prefixes
 # Manifolds: Identity, TranslationGroup
+#
 include("documentation_glossary.jl")
 include("utils.jl")
 include("interface.jl")
@@ -157,7 +159,8 @@ export hat, hat!
 export inv, inv!, inv_left_compose, inv_left_compose!, inv_right_compose, inv_right_compose!
 export isapprox, is_point, is_vector
 export exp, exp!
-export identity_element, identity_element!, is_identity, inv, inv!, diff_inv, diff_inv!
+export identity_element, identity_element!, is_identity, inner
+export inv, inv!, diff_inv, diff_inv!
 export jacobian_conjugate, jacobian_conjugate!
 export lie_bracket, lie_bracket!, log, log!
 export manifold_dimension