+++ src/interface.jl
@@ -490,6 +490,8 @@ function identity_element! end
 @doc "$(_doc_identity_element)"
 identity_element!(G::AbstractLieGroup, e)
 
+ManifoldsBase.inner(G::AbstractLieGroup, g, X, Y) = inner(LieAlgebra(G), X, Y)
+
 _doc_inv = """
     inv(G::AbstractLieGroup, g)
     inv!(G::AbstractLieGroup, h, g)
@@ -809,6 +811,8 @@ function ManifoldsBase.log!(
 end
 
 ManifoldsBase.manifold_dimension(G::AbstractLieGroup) = manifold_dimension(base_manifold(G))
+
+ManifoldsBase.norm(G::AbstractLieGroup, g, X) = norm(LieAlgebra(G), X)
 
 _doc_rand = """
     rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...)